backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6

# AI & Services
openai>=1.30.0
easyocr>=1.7.1
python-dotenv==1.0.0